Paginate the config Webpages
The configuration pages got too large to be reliably being
constructed in a single String() - the page is meanwhile almost
10kb which is dangerously close to the max heap available.

We can paginate the output of the page in chunks of ~ 2000 bytes
which reduces memory pressure while not causing higher packet
overhead.

Unfortunately calling server.sendContent() is too much of a code
bloat for each line, so we need to construct semi-large
strings and then send them out.